refactor: extract get_project_root to utils module
  - Create src/wct/utils.py for shared utility functions
  - Move get_project_root from logging.py to utils.py for better separation of concerns
  - Update imports in logging.py and test_logging.py
  - Add config/organisation.yaml for GDPR Article 30(1)(a) compliance preparation
