+++ src/wct/logging.py
@@ -6,7 +6,6 @@ using logging.config.dictConfig() with YAML configuration files stored in src/wc
 
 from __future__ import annotations
 
-import importlib.util
 import logging
 import logging.config
 import os
@@ -16,74 +15,13 @@ from typing import Any, cast
 
 import yaml
 
+from wct.utils import ProjectUtilsError, get_project_root
+
 
 class LoggingError(Exception):
     """Exception raised for logging configuration errors."""
 
     pass
-
-
-def get_project_root() -> Path:
-    """Get the project root directory using robust discovery methods.
-
-    This function follows industry-standard approaches used by tools like pytest,
-    black, and setuptools to find the project root:
-    1. Package location via importlib (handles installed packages)
-    2. Upward search for project markers (handles development)
-
-    Returns:
-        Path to the project root directory
-
-    Raises:
-        LoggingError: If project root cannot be determined
-
-    """
-    # Method 1: Use package location if installed (preferred)
-    try:
-        spec = importlib.util.find_spec("wct")
-        if spec and spec.origin:
-            package_path = Path(spec.origin).parent
-
-            # For development installs, the package might be a symlink to src/wct
-            if package_path.is_symlink():
-                real_package = package_path.resolve()
-                if real_package.name == "wct" and real_package.parent.name == "src":
-                    potential_root = real_package.parent.parent
-                    if (potential_root / "src" / "wct" / "config").is_dir():
-                        return potential_root
-
-            # For regular installs, search upward from package location
-            for path in [package_path, *package_path.parents]:
-                if (path / "src" / "wct" / "config").is_dir():
-                    return path
-    except (ImportError, AttributeError):
-        # Package not found via importlib, continue to marker-based discovery
-        pass
-
-    # Method 2: Search upward from current file for project markers (development)
-    current_path = Path(__file__).parent.resolve()
-
-    for path in [current_path, *current_path.parents]:
-        # Look for common project root markers
-        markers = [
-            path / "pyproject.toml",  # Modern Python project file
-            path / "setup.py",  # Legacy setup file
-            path / ".git",  # Git repository root
-            path / "src" / "wct",  # Our package structure
-        ]
-
-        if any(marker.exists() for marker in markers):
-            # Verify this looks like our project by checking for src/wct/config/
-            if (path / "src" / "wct" / "config").is_dir():
-                return path
-
-    # If both methods fail, provide clear error message
-    raise LoggingError(
-        f"Could not determine project root. Searched:\n"
-        f"1. Package location via importlib.util.find_spec('wct')\n"
-        f"2. Upward from {current_path} for project markers\n"
-        f"Expected to find a directory containing 'src/wct/config/' subdirectory."
-    )
 
 
 def get_config_path(
@@ -259,7 +197,7 @@ def setup_logging(
             "Logging configured from: %s", config_path.relative_to(get_project_root())
         )
 
-    except (LoggingError, ImportError, KeyError, ValueError) as e:
+    except (LoggingError, ProjectUtilsError, ImportError, KeyError, ValueError) as e:
         # Fallback to basic logging on any configuration error
         fallback_level = level or "INFO"
         _setup_basic_logging(fallback_level)
